Split Segments and Lists into two different pages
Before both Segments and Lists where acessible via two different tabs on
the page `/wp-admin/admin.php?page=mailpoet-segments`. Now they are
separated into their own pages:
`/wp-admin/admin.php?page=mailpoet-segments` and
`/wp-admin/admin.php?page=mailpoet-lists`.

[MAILPOET-5392]
