+++ mailpoet/assets/js/src/segments/heading.tsx
@@ -6,22 +6,26 @@ import { SubscribersInPlan } from 'common/subscribers_in_plan';
 import { MssAccessNotices } from 'notices/mss_access_notices';
 import { SubscribersCacheMessage } from 'common/subscribers_cache_message';
 
-function ListHeading(): JSX.Element {
+function ListHeading({ segmentType }): JSX.Element {
   return (
     <>
       <TopBarWithBeamer>
-        <Link className="mailpoet-button button-secondary" to="/new">
-          {plusIcon}
-          <span>{MailPoet.I18n.t('new')}</span>
-        </Link>
-        <Link
-          className="mailpoet-button button-secondary"
-          to="/new-segment"
-          data-automation-id="new-segment"
-        >
-          {plusIcon}
-          <span>{MailPoet.I18n.t('newSegment')}</span>
-        </Link>
+        {segmentType === 'static' && (
+          <Link className="mailpoet-button button-secondary" to="/new">
+            {plusIcon}
+            <span>{MailPoet.I18n.t('new')}</span>
+          </Link>
+        )}
+        {segmentType === 'dynamic' && (
+          <Link
+            className="mailpoet-button button-secondary"
+            to="/new-segment"
+            data-automation-id="new-segment"
+          >
+            {plusIcon}
+            <span>{MailPoet.I18n.t('newSegment')}</span>
+          </Link>
+        )}
       </TopBarWithBeamer>
 
       <SubscribersInPlan
