+++ mailpoet/assets/js/src/segments/static/form.tsx
